+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+* [v1.16.0](https://github.com/newt-sc/a4kStreaming/releases/tag/plugin.video.a4kstreaming%2Fplugin.video.a4kstreaming-1.16.0):
+  * New settings:
+    * Autoplay - instead of showing source select, auto choose the first source that resolves
+  * New context menu:
+    * Force source select - only available when Autoplay is enabled
+  * Add more episode auto selection checks
+  * Allow browsing cast and crew to find titles which they are known for. (This is new API which currently returns test results)
+
 * [v1.15.0](https://github.com/newt-sc/a4kStreaming/releases/tag/plugin.video.a4kstreaming%2Fplugin.video.a4kstreaming-1.15.0):
   * Add RD expired notification